Rename mockTimeEntry to cartesianChartInput
